Update 28 sep 2019 23:52

+++ application/modules/student/controllers/Student.php
@@ -322,7 +322,9 @@ function do_upload() {
         //
       $result = $file_nm;
       $lspdf = $this->input->post('last-pdf');
+      if ($lspdf != "") {
       unlink("./assets/ijazah/$lspdf");
+      }
     }else{
       $result = $this->input->post('last-pdf');
     }
